pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "img-timeline"
version = "0.1.0"
description = "Create color timeline images from TIFF frame sequences."
readme = "README.md"
requires-python = ">=3.10"
license = { file = "LICENSE" }
authors = [{ name = "aelder" }]
dependencies = [
  "Pillow>=10.0.0",
]

[project.optional-dependencies]
progress = ["tqdm>=4.66.0"]
dev = [
  "ruff>=0.6.0",
]

[project.scripts]
img-timeline = "img_timeline.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
